+++ locale/zh-cn/script.lua
@@ -128,6 +128,10 @@ DIAG_UNKNOWN_CAST_VARIABLE            =
 '未知的类型转换变量 `{}`。'
 DIAG_CAST_TYPE_MISMATCH               =
 '不能将 `{ref}` 转换为 `{def}`。'
+DIAG_MISSING_RETURN_VALUE             =
+'至少需要 {min} 个返回值，但此处只返回 {rmax} 个值。'
+DIAG_MISSING_RETURN_VALUE             =
+'至少需要 {min} 个返回值，但此处只返回 {rmin} 到 {rmax} 个值。'
 
 MWS_NOT_SUPPORT         =
 '{} 目前还不支持多工作目录，我可能需要重启才能支持新的工作目录...'