+++ script/parser/guide.lua
@@ -12,7 +12,7 @@ local type         = type
 ---@field tag                   string
 ---@field args                  { [integer]: parser.object, start: integer, finish: integer }
 ---@field locals                parser.object[]
----@field returns               parser.object[]
+---@field returns?              parser.object[]
 ---@field exps                  parser.object[]
 ---@field keys                  parser.object[]
 ---@field uri                   uri